apps/haim-auth/requirements.txt
fastapi-jwt-auth
psycopg2-binary
python-dotenv
fastapi-jwt
websockets
uvicorn
fastapi
bcrypt